cash defi
As a decentralized financial system on top of the Bitcoin Cash blockchain, our coin aims to enable users to carry out financial transactions such as lending, leverage trading and staking in a simple and secure manner. It is also intended to be part of a broader ecosystem of applications and services built on top of the Bitcoin Cash blockchain. Our token also offers a capped total supply and regular halvings every 4 years, limiting supply and incentivizing users to hold it and use it for the long term.